Adding confirm button

+++ app/Controllers/ListController.js
@@ -2,6 +2,8 @@ import ListService from "../Services/ListService.js";
 import _store from "../store.js"
 import List from "../Models/List.js";
 
+
+
 //TODO Don't forget to render to the screen after every data change.
 function _drawLists() {
   let template = ''
@@ -30,8 +32,25 @@ export default class ListController {
 
 
   deleteList(listId) { /***************DELETE LIST***************/
-    ListService.deleteList(listId)
-    _drawLists()
+    Swal.fire({
+      title: 'Are you sure?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        Swal.fire({
+          position: 'center',
+          title: 'Deleted!',
+          showConfirmButton: false,
+          timer: 500
+        })
+        ListService.deleteList(listId)
+        _drawLists()
+      }
+    })
   }
 
   addTask(event, listId) { /***************ADD TASK***************/
@@ -40,9 +59,26 @@ export default class ListController {
     _drawLists()
   }
 
-  deleteTask(listId) { /***************DELETE TASK***************/
-    ListService.deleteTask(listId)
-    _drawLists()
+  deleteTask(listId, taskIndex) { /***************DELETE TASK***************/
+    Swal.fire({
+      title: 'Are you sure?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        Swal.fire({
+          position: 'center',
+          title: 'Deleted!',
+          showConfirmButton: false,
+          timer: 500
+        })
+        ListService.deleteTask(listId, taskIndex)
+        _drawLists()
+      }
+    })
   }
 
   //TODO: Your app will need the ability to create, and delete both lists and listItems
